new recipes added to seeder

+++ database/migrations/2021_06_30_115744_create_ingredient_recipe_table.php
@@ -17,7 +17,7 @@ class CreateIngredientRecipeTable extends Migration
             $table->id();
             $table->foreignId('ingredient_id');
             $table->foreignId('recipe_id');
-            $table->integer('quantity')->nullable();
+            $table->float('quantity')->nullable();
             $table->string('measurement')->nullable();
             $table->timestamps();
         });
